requirements.txt
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
glcontext==3.0.0
glfw==2.7.0
imgui==2.0.0
Jinja2==3.1.4
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
moderngl==5.12.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyassimp==5.2.5
pyglet==1.5.29
PyGLM==2.7.2
PyOpenGL==3.1.1a1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavefront==1.3.3
scipy==1.14.1
setuptools==72.1.0
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformations==2024.5.24
trimesh==4.4.9
typing_extensions==4.12.2
tzdata==2024.1
wheel==0.43.0
xatlas==0.0.9
